requirements.txt

azure-ai-evaluation
azure-identity
openai
python-dotenv
rich